cp before refactoring to myExperiments
