+++ app/src/main/java/com/example/appraisal/UI/MainActivity.java
@@ -6,7 +6,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import com.example.appraisal.R;
-import com.example.appraisal.UI.subscription.ExpSubscriptionActivity;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -19,6 +18,12 @@ public class MainActivity extends AppCompatActivity {
     public void toSelection(View v) {
         // TODO Bring the user from main menu, to experiment type menu
         Intent intent = new Intent(this, ExpSubscriptionActivity.class);
+        startActivity(intent);
+    }
+
+    public void toMyExperiments(View v){
+        // TODO Bring the user from main menu, to experiment type menu
+        Intent intent = new Intent(this, MyExperimentActivity.class);
         startActivity(intent);
     }
 }
